[TS-30] Refactor: Habit tip refactoring
