+++ src/components/Home/HabitGenerate/HabitGenerate.style.ts
@@ -37,48 +37,6 @@ export const profileBoxStyle = css`
 	}
 `;
 
-export const tipBoxStyle = css`
-	position: relative;
-	${theme.font.body_b};
-
-	& > button {
-		align-items: center;
-		padding: 0.5rem 1rem;
-		border-radius: 20px;
-		color: white;
-		background-color: ${theme.color.main_deep_blue};
-	}
-
-	& > div {
-		transform: translateY(calc(12px));
-		position: absolute;
-		display: flex;
-		flex-direction: column;
-		background-color: white;
-		white-space: pre-wrap;
-		padding: 1rem 3.5rem 1rem 1rem;
-		border: 1px solid ${theme.color.main_blue};
-		border-radius: 5px;
-		gap: 20px;
-
-		& > div {
-			width: 14.75rem;
-
-			& > p:first-of-type {
-				${theme.font.head_c}
-				color: ${theme.color.main_blue}
-			}
-		}
-
-		& > svg {
-			position: absolute;
-			transform: translate(-1rem, 1rem);
-			top: 0;
-			right: 0;
-		}
-	}
-`;
-
 export const selectBoxStyle = css`
 	display: flex;
 	flex-direction: column;
